dunfell: Remove untested/unsupported boards from manifest
As we move to a CI workflow, currently only those machines are tested
for the release. Those variants are integrated in the CI-Testing loop
and will be fully tested with daily fresh stable patches. During the
PD21.1 development cycle more SoM variants will be added. Minor release
will be made when there are no issues in the daily test results.

Signed-off-by: Yunus Bas <y.bas@phytec.de>
Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>